+++ gamemodes/terrortown/gamemode/shared/lang/de.lua
@@ -1349,3 +1349,18 @@ L.karma_enemyhurt_tooltip = "Schaden"
 L.karma_cleanround_tooltip = "Saubere Runde"
 L.karma_roundheal_tooltip = "Rundenheilung"
 L.karma_unknown_tooltip = "Unbekannt"
+
+-- 2021-05-07
+L.header_random_shop_administration = "Zufalls-Shop"
+L.header_random_shop_value_administration = "Balance Einstellungen"
+
+L.shopeditor_name_random_shops = "Aktiviere Zufalls-Shop"
+L.shopeditor_desc_random_shops = [[Zufalls-Shops geben jedem Spieler eine zufällige Menge an Ausrüstung.
+Team-Shops zwingen alle Spieler aus einem Team zu ein und demselben Shopinhalt.
+Auswürfeln erlaubt es ein gegebenes Set für Credits durch ein neues Zufallsset auszutauschen.]]
+L.shopeditor_name_random_shop_items = "Anzahl zufälliger Equipments"
+L.shopeditor_desc_random_shop_items = "Dies enthält Equipments, welche mit .noRandom (immer enthalten) markiert sind. Also wähle die Anzahl hoch genug, damit nicht nur solches Equipment im Shop landet."
+L.shopeditor_name_random_team_shops = "Aktiviere Team-Shops"
+L.shopeditor_name_random_shop_reroll = "Aktiviere Möglichkeit Shop neu auszuwürfeln"
+L.shopeditor_name_random_shop_reroll_cost = "Kosten pro Auswürfeln"
+L.shopeditor_name_random_shop_reroll_per_buy = "Automatisches Auswürfeln nach jedem Kauf"